style: rename dirs
